+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Unreleased
 ==========
 
 * Update README.
+* Allow user to specify a list of paths that s3:getObject is set on rather than
+  doing it on the whole bucket (#4).
 
 0.1a0 - 20th June 2018
 ======================